add: First tests
