+++ src/Components/CalendarCard/CalendarCard.module.css
@@ -1,0 +1,58 @@
+.cardAppointment{
+    margin-bottom: 20px;
+    position: relative;
+}
+
+.month{
+    position: absolute;
+    top:-5px;
+    right: -5px;
+    background: #0a0a0a;
+    color: #efefef;
+    padding: 10px;
+    font-size: 14px;
+    border-radius: 0 0 0 10px;
+}
+
+.day{
+    font-size: 4rem;
+    text-align: center;
+    width:100%;
+    display: block;
+    background: #cf000f;
+    color: #fff
+}
+
+.nomeHolder{
+    display: flex;
+    justify-content: space-between;
+    align-items: stretch;
+    gap: 10px;
+    padding: 10px 10px;
+}
+
+.nomePaciente, .nomeDentista{
+    margin-bottom: 0;
+}
+
+.verticalLine{
+    width:1px;
+    background: #dfdfdf
+}
+
+.whoIs{
+    font-size:0.6rem;
+    display: block;
+    text-transform: uppercase;
+}
+
+.deleteButton{
+    position: absolute;
+    top:-5px;
+    left: -5px;
+    background: #0a0a0a;
+    color: #efefef;
+    padding: 10px;
+    border-radius: 0 0 10px 0;
+    margin: 0;
+}
