+++ src/setupTests.js
@@ -3,16 +3,16 @@
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
 import '@testing-library/jest-dom';
-import { server } from './mocks/server.js'
+// qimport { server } from './mocks/server.js'
 
 // Establish API mocking before all tests.
-beforeAll(() => server.listen())
+// beforeAll(() => server.listen())
 
-// Reset any request handlers that we may add during the tests,
-// so they don't affect other tests.
-afterEach(() => server.resetHandlers())
+// // Reset any request handlers that we may add during the tests,
+// // so they don't affect other tests.
+// afterEach(() => server.resetHandlers())
 
-// Clean up after the tests are finished.
-afterAll(() => server.close())
+// // Clean up after the tests are finished.
+// afterAll(() => server.close())
 
 window.alert = ()=>{}
